+++ ImageD11src/grid_index_parallel.py
@@ -104,7 +104,7 @@ def doindex( gve, x, y, z, w, gridpars):
         )
     # added in indexer.__init__
     #myindexer.ds = np.sqrt( (gve * gve).sum(axis=0) )
-    #myindexer.ga = np.zeros(len(myindexer.ds),np.int)-1 # Grain assignments
+    #myindexer.ga = np.zeros(len(myindexer.ds),int)-1 # Grain assignments
     for ring1 in gridpars['RING1']:
         for ring2 in gridpars['RING2']:
             myindexer.parameterobj.set_parameters(  {
@@ -224,10 +224,10 @@ def initgrid( fltfile, parfile, tmp, gridpars ):
     gridpars[ 'UC' ] = unitcell.unitcell_from_parameters( mytransformer.parameterobj )
     col = mytransformer.colfile
     if not "drlv2" in col.titles:
-        col.addcolumn( np.ones(col.nrows, np.float),
+        col.addcolumn( np.ones(col.nrows, float),
                        "drlv2" )
     if not "labels" in col.titles:
-        col.addcolumn( np.ones(col.nrows, np.float)-2,
+        col.addcolumn( np.ones(col.nrows, float)-2,
                        "labels" )
     if not "sc" in col.titles:
         assert "xc" in col.titles
